requirements.txt
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
EbookLib==0.18
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
Markdown==3.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.2.1
